Break up the BagsApi feature test to reflect the lookup traits
